src/pyproject.toml
[tool.poetry]
name = "bajkal"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
amqp = "^5.3.1"
astroid = "^3.3.6"
beautifulsoup4 = "^4.12.3"
billiard = "^4.2.1"
boto3 = "^1.35.84"
celery = "^5.4.0"
certifi = "2023.11.17"
cffi = "^1.17.1"
charset-normalizer = "^3.4.0"
click = "^8.1.7"
click-didyoumean = "^0.3.1"
click-plugins = "^1.1.1"
click-repl = "^0.3.0"
colorama = "^0.4.6"
colorlog = "^6.9.0"
cron-descriptor = "^1.4.5"
cryptography = "^44.0.0"
cssbeautifier = "^1.15.1"
defusedxml = "^0.7.1"
dill = "^0.3.9"
dj-database-url = "^2.3.0"
dj-stripe = "^2.8.4"
django = "^5.1.4"
django-allauth = "^65.3.0"
django-appconf = "^1.0.6"
django-celery-beat = "^2.7.0"
django-cleanup = "8.0.0"
django-clearcache = "^1.2.1"
django-cprofile-middleware = "^1.0.5"
django-extensions = "^3.2.3"
django-htmx = "^1.21.0"
django-jazzmin = "^3.0.1"
django-polymorphic = "^3.1.0"
django-redis = "^5.4.0"
django-seed = "^0.3.1"
django-silk = "^5.3.2"
django-split-settings = "^1.3.2"
django-storages = "^1.14.4"
django-timezone-field = "^7.0"
django-use-email-as-username = "^1.4.0"
editorconfig = "^0.17.0"
faker = "^33.1.0"
fontawesomefree = "^6.6.0"
gprof2dot = "^2024.6.6"
gunicorn = "^23.0.0"
html-tag-names = "^0.1.2"
html-void-elements = "^0.1.0"
idna = "^3.10"
iniconfig = "^2.0.0"
isort = "^5.13.2"
jmespath = "^1.0.1"
jsbeautifier = "^1.15.1"
json5 = "^0.10.0"
kombu = "^5.4.2"
lazy-object-proxy = "^1.10.0"
libsass = "^0.23.0"
markdown = "^3.7"
mccabe = "^0.7.0"
mypy-extensions = "^1.0.0"
oauthlib = "^3.2.2"
packaging = "^24.2"
pathspec = "^0.12.1"
pillow = "^11.0.0"
platformdirs = "^4.3.6"
plotly = "^5.24.1"
pluggy = "^1.5.0"
prompt-toolkit = "^3.0.48"
psycopg2 = "^2.9.10"
pycparser = "^2.22"
pyjwt = "^2.10.1"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
python-crontab = "^3.2.0"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.0.1"
python3-openid = "^3.2.0"
pyyaml = "^6.0.2"
redis = "^5.2.1"
regex = "^2024.11.6"
requests = "^2.32.3"
requests-oauthlib = "^2.0.0"
s3transfer = "^0.10.4"
six = "^1.17.0"
soupsieve = "^2.6"
sqlparse = "^0.5.3"
tenacity = "^9.0.0"
tomlkit = "^0.13.2"
toposort = "^1.10"
tqdm = "^4.67.1"
typing-extensions = "^4.12.2"
tzdata = "^2024.2"
urllib3 = "^2.2.3"
vine = "^5.1.0"
wcwidth = "^0.2.13"
wrapt = "^1.17.0"
stripe = "^11.4.0"
setuptools = "^75.6.0"
coverage = "^7.6.9"
pytest-cov = "^6.0.0"
unidecode = "^1.3.8"
matplotlib = "^3.10.0"
numpy = "^2.2.2"
pandas = "^2.2.3"
calplot = "^0.1.7.5"
mpld3 = "^0.5.10"

[tool.poetry.group.dev.dependencies]
pre-commit = "4.0.1"
nplusone = "1.0.0"
django-debug-toolbar = "^4.4.6"
djlint = "^1.36.4"
vulture = "^2.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.djlint]
profile = "django"
